Created RSSFeedHelpers, moved the helper functions there
